requirements.txt
Flask==3.0.0
Pillow==10.1.0
werkzeug==3.0.1
cloudinary==1.36.0
python-dotenv==1.0.0
requests==2.31.0
sentry-sdk[flask]==1.40.6